@@ -2,28 +2,26 @@ package dev.kylejulian.twsmanagement.data;
 
 import java.io.IOException;
 import java.net.HttpURLConnection;
+import java.net.URI;
 import java.net.URL;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dev.kylejulian.twsmanagement.data.entities.MojangUserModel;
 import dev.kylejulian.twsmanagement.data.entities.MojangUserNameModel;
+import dev.kylejulian.twsmanagement.util.LogUtils;
 import net.kyori.adventure.text.Component;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 public class MojangApi {
 
-	private final Logger logger;
 	private final Map<String, UUID> cachedIds;
 
-	public MojangApi(Logger logger) {
-		this.logger = logger;
+	public MojangApi() {
 		this.cachedIds = new HashMap<>();
 	}
 
@@ -32,14 +30,14 @@ public class MojangApi {
 		return CompletableFuture.supplyAsync(() -> {
 			try {
 				if (this.cachedIds.containsKey(name)) {
-					this.logger.log(Level.FINE, "[MojangApi] Cache hit");
+					LogUtils.debug("[MojangApi] Cache hit");
 					return this.cachedIds.get(name);
 				}
 
-				this.logger.log(Level.FINE, "[MojangApi] Cache miss");
+				LogUtils.debug("[MojangApi] Cache miss");
 
 				String url = "https://api.mojang.com/users/profiles/minecraft/" + name;
-				URL obj = new URL(url);
+				URL obj = URI.create(url).toURL();
 				HttpURLConnection con = (HttpURLConnection) obj.openConnection();
 
 				con.setRequestMethod("GET");
@@ -54,7 +52,7 @@ public class MojangApi {
 				return userId;
 			}
 			catch (IOException e) {
-				this.logger.log(Level.INFO, "Unable to find Player {0} from the Mojang Api", name);
+				LogUtils.warn("Unable to find Player "+ name + " from the Mojang Api");
 			}
 
 			return null;
@@ -68,19 +66,19 @@ public class MojangApi {
 				if (this.cachedIds.containsValue(playerId)) {
 					for (Map.Entry<String, UUID> entry : this.cachedIds.entrySet()) {
 						if (entry.getValue().equals(playerId)) {
-							this.logger.log(Level.FINE, "[MojangApi] Cache hit");
+							LogUtils.debug("[MojangApi] Cache hit");
 							return Component.text(entry.getKey());
 						}
 					}
 				}
 
-				this.logger.log(Level.FINE, "[MojangApi] Cache miss");
+				LogUtils.debug("[MojangApi] Cache miss");
 
 				String playerIdString = playerId.toString();
 
 				String url = "https://api.mojang.com/user/profiles/" +
 						playerIdString.replace("-", "") + "/names";
-				URL obj = new URL(url);
+				URL obj = URI.create(url).toURL();
 				HttpURLConnection con = (HttpURLConnection) obj.openConnection();
 
 				ObjectMapper mapper = new ObjectMapper();
@@ -95,11 +93,12 @@ public class MojangApi {
 				return Component.text(name);
 			}
 			catch (IOException e) {
-				this.logger.log(Level.WARNING, "Unable to find PlayerId {0} from the Mojang Api", playerId);
+				LogUtils.warn("Unable to find PlayerId " + playerId + " from the Mojang Api");
+				LogUtils.warn(e.getMessage());
 			}
 			catch (Exception e) {
-				this.logger.log(Level.SEVERE, "Unable to find PlayerId {0} from the Mojang Api", playerId);
-				e.printStackTrace();
+				LogUtils.error("Unable to find PlayerId " + playerId + " from the Mojang Api");
+				LogUtils.error(e.getMessage());
 			}
 
 			return null;
